django/requirements.txt
addict==2.1.3
argcomplete==1.8.2
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.5.3
boto==2.49.0
boto3==1.7.50
botocore==1.10.50
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
commonmark==0.9.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
diff-match-patch==20121119
Django==2.0.7
django-cors-headers==2.3.0
django-grappelli==2.12.1
django-import-export==1.0.1
django-jet==1.0.7
django-multi-email-field==0.5.1
django-pagedown==1.0.6
django-rest-swagger==2.2.0
django-ses==0.8.7
django-storages==1.6.6
django-suit==0.2.26
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
docutils==0.14
docx2txt==0.6
EbookLib==0.15
et-xmlfile==1.0.1
Fabric3==1.14.post1
feedparser==5.2.1
future==0.17.1
geographiclib==1.49
geopy==1.17.0
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.3
lxml==4.2.5
m2r==0.2.1
Markdown==2.6.11
markdown2==2.3.7
markdownify==0.4.1
MarkupSafe==1.0
martor==1.2.7
mistune==0.8.4
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.7
paramiko==2.4.1
Pillow==5.2.0
psycopg2-binary==2.7.5
pyasn1==0.4.3
pycparser==2.18
PyJWT==1.6.4
PyNaCl==1.2.1
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
tomd==0.1.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.12.0
